+++ src/ldc/pretrain/_jsonlines.py
@@ -1,10 +1,11 @@
 import argparse
 import jsonlines
+import os
 from typing import Iterable, List, Union
 
 from ldc.core import LOGGING_WARN, domain_suffix
-from ldc.io import locate_files, open_file, generate_output
-from ._core import PretrainData, PretrainReader, BatchPretrainWriter
+from ldc.io import locate_files, open_file, generate_output, is_compressed
+from ._core import PretrainData, PretrainReader, StreamPretrainWriter
 from ldc.utils import add_meta_data
 
 
@@ -160,12 +161,13 @@ class JsonLinesPretrainReader(PretrainReader):
             self._current_input = None
 
 
-class JsonLinesPretrainWriter(BatchPretrainWriter):
+class JsonLinesPretrainWriter(StreamPretrainWriter):
     """
     Writer for the JsonLines JSON format.
     """
 
     def __init__(self, target: str = None, att_content: str = None, att_id: str = None,
+                 num_digits: int = 6, buffer_size: int = 1000,
                  logger_name: str = None, logging_level: str = LOGGING_WARN):
         """
         Initializes the writer.
@@ -176,6 +178,10 @@ class JsonLinesPretrainWriter(BatchPretrainWriter):
         :type att_content: str
         :param att_id: the (optional) attribute with the ID
         :type att_id: str
+        :param num_digits: the number of digits to use for the output file names
+        :type num_digits: int
+        :param buffer_size: the size of the record buffer (< 1 for unlimited)
+        :type buffer_size: int
         :param logger_name: the name to use for the logger
         :type logger_name: str
         :param logging_level: the logging level to use
@@ -185,9 +191,12 @@ class JsonLinesPretrainWriter(BatchPretrainWriter):
         self.target = target
         self.att_content = att_content
         self.att_id = att_id
-        self._current_output = None
-        self._output = None
-        self._writer = None
+        self.num_digits = num_digits
+        self.buffer_size = buffer_size
+        self._concatenate = False
+        self._first_item = True
+        self._fname_format = None
+        self._buffer = []
 
     def name(self) -> str:
         """
@@ -218,6 +227,8 @@ class JsonLinesPretrainWriter(BatchPretrainWriter):
         parser.add_argument("-o", "--output", type=str, help="Path of the JsonLines file to write (directory when processing multiple files)", required=True)
         parser.add_argument("--att_content", metavar="ATT", type=str, default=None, help="The attribute for the text content", required=False)
         parser.add_argument("--att_id", metavar="ATT", type=str, default=None, help="The name of the attribute for the row IDs (uses 'id' from meta-data)", required=False)
+        parser.add_argument("-d", "--num_digits", metavar="NUM", type=int, default=6, help="The number of digits to use for the filenames", required=False)
+        parser.add_argument("-b", "--buffer_size", metavar="SIZE", type=int, default=1000, help="The size of the record buffer when concatenating (to improve I/O throughput)", required=False)
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -231,6 +242,8 @@ class JsonLinesPretrainWriter(BatchPretrainWriter):
         self.target = ns.output
         self.att_content = ns.att_content
         self.att_id = ns.att_id
+        self.num_digits = ns.num_digits
+        self.buffer_size = ns.buffer_size
 
     def initialize(self):
         """
@@ -239,35 +252,85 @@ class JsonLinesPretrainWriter(BatchPretrainWriter):
         super().initialize()
         if self.att_content is None:
             raise Exception("No content attribute specified!")
+        self._first_item = True
+        self._fname_format = "%0" + str(self.num_digits) + "d.txt"
+        if os.path.exists(self.target) and os.path.isdir(self.target):
+            self._concatenate = False
+        else:
+            self._concatenate = True
+            if is_compressed(self.target):
+                raise Exception("Cannot use compression when concatenating due to streaming!")
+        self._buffer.clear()
 
-    def write_batch(self, data: Iterable[PretrainData]):
+    def _write(self, data: List[PretrainData], output: str, mode: str):
         """
-        Saves the data in one go.
+        Writes the data to disk.
 
-        :param data: the data to write as iterable of PretrainData
-        :type data: Iterable
+        :param data: the records to write
+        :type data: list
+        :param output: the file to write to
+        :type output: str
+        :param mode: the file mode to use
+        :type mode: str
         """
-        if self._has_input_changed(update=True) and self._output_needs_changing(self._current_output, self.target, ".jsonl"):
-            self.finalize()
-            self._current_output = generate_output(self.session.current_input, self.target, ".jsonl", self.session.options.compression)
-            self.logger().info("Writing to: " + self._current_output)
-            self._output = open_file(self._current_output, mode="wt")
-            self._writer = jsonlines.Writer(self._output)
+        with open(output, mode) as fp:
+            writer = jsonlines.Writer(fp)
+            for item in data:
+                d = {self.att_content: item.content}
+                if self.att_id is not None:
+                    if (item.meta is not None) and ("id" in item.meta):
+                        d[self.att_id] = item.meta["id"]
+                try:
+                    writer.write(d)
+                except KeyboardInterrupt as e:
+                    raise e
+                except:
+                    self.logger().exception("Failed to write record: %s" % str(d))
+            writer.close()
 
-        for item in data:
-            d = {self.att_content: item.content}
-            if self.att_id is not None:
+    def _flush_buffer(self):
+        """
+        Writes the buffer content to disk.
+        """
+        self.logger().debug("flushing buffer: %d" % len(self._buffer))
+        mode = "w" if self._first_item else "a"
+        if self._first_item:
+            self.logger().info("Writing to: %s" % self.target)
+        self._first_item = False
+        self._write(self._buffer, self.target, mode)
+        self._buffer.clear()
+
+    def write_stream(self, data: Union[PretrainData, Iterable[PretrainData]]):
+        """
+        Saves the data one by one.
+
+        :param data: the data to write
+        :type data: PretrainData
+        """
+        if isinstance(data, PretrainData):
+            data = [data]
+
+        if self._concatenate:
+            self._buffer.extend(data)
+            if len(self._buffer) >= self.buffer_size:
+                self._flush_buffer()
+        else:
+            for item in data:
                 if (item.meta is not None) and ("id" in item.meta):
-                    d[self.att_id] = item.meta["id"]
-            self._writer.write(d)
+                    try:
+                        fname = self._fname_format % int(item.meta["id"])
+                    except:
+                        fname = str(item.meta["id"]) + ".jsonl"
+                else:
+                    fname = self._fname_format % self.session.count
+                output = generate_output(fname, self.target, ".jsonl", self.session.options.compression)
+                self.logger().info("Writing to: %s" % output)
+                self._write([item], output, "w")
 
     def finalize(self):
         """
-        Finishes the writing, e.g., for closing files or databases.
+        Finishes the processing, e.g., for closing files or databases.
         """
-        if self._output is not None:
-            super().finalize()
-            self._writer.close()
-            self._writer = None
-            self._output.close()
-            self._output = None
+        super().finalize()
+        if len(self._buffer) > 0:
+            self._flush_buffer()